migrating to postgresql'

+++ src/main/resources/schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS Run (
     title varchar(250) NOT NULL,
     started_on timestamp NOT NULL,
     completed_on timestamp NOT NULL,
-    distance DOUBLE NOT NULL,
+    distance DOUBLE PRECISION NOT NULL,
     location varchar(10) NOT NULL,
     PRIMARY KEY (id)
 );
